feat(miner-governor): build a real production runSlopAssessment implementation (#5133)
attempt-runner.js's deps.runSlopAssessment had no production binding, only
a test double. Extracts the PR-side deterministic slop scorer from
src/signals/slop.ts to packages/gittensory-engine/src/signals/slop.ts
(byte-parity verified: all 63 pre-existing tests pass unchanged, and every
extracted function diffed identical against the original), so the miner
CLI can run the SAME scorer the live gate uses. src/signals/slop.ts
becomes a partial shim re-exporting the PR-side pieces; issue-side triage
stays local (not needed by the miner). Wires the now-portable
buildSlopAssessment into a real runSlopAssessment binding in
packages/gittensory-miner, matching self-review-adapter.ts's own
structural-mirror design.
