+++ packages/gittensory-miner/lib/slop-assessment.d.ts
@@ -1,0 +1,3 @@
+import type { SlopAssessment, SlopAssessmentInput } from "@jsonbored/gittensory-engine";
+
+export function runSlopAssessment(input: SlopAssessmentInput): SlopAssessment;
